backend/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
google-generativeai>=0.8.0
google-genai>=1.0.0
python-dotenv==1.2.2
pydantic==2.13.3
langchain>=1.0.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
chromadb>=1.0.0
sse-starlette>=1.6.0
